services/ingestion-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
pydantic
psycopg2-binary
sqlalchemy
alembic
PyPDF2
python-docx
pymupdf
